Feat: useSlider
